+++ src/ultimate_notion/obj_api/core.py
@@ -77,7 +77,7 @@ BLOCK_URL_LONG_RE = re.compile(
 )
 
 
-T = TypeVar('T')  # ToDo: use new syntax in Python 3.12 and consider using default = in Python 3.13+
+T = TypeVar('T', default=Any)  # ToDo: use new syntax in Python 3.12 and consider using default = in Python 3.13+
 
 
 # ToDo: Replace with MISSING when stable: https://docs.pydantic.dev/latest/concepts/experimental/#missing-sentinel