+++ tests/test_blocks.py
@@ -12,7 +12,7 @@ from ultimate_notion.blocks import ChildrenMixin
 from ultimate_notion.errors import InvalidAPIUsageError
 from ultimate_notion.obj_api import objects as objs
 from ultimate_notion.obj_api.blocks import Block as ObjBlock
-from ultimate_notion.obj_api.core import Unset, UserRef
+from ultimate_notion.obj_api.core import ObjectRef, Unset, UserRef
 from ultimate_notion.rich_text import Text
 
 
@@ -866,9 +866,27 @@ def test_local_remote_text_equality(notion: uno.Session, root_page: uno.Page) ->
 
 @pytest.mark.vcr()
 def test_local_remote_mention_block_cmp(root_page: uno.Page, notion: uno.Session, person: uno.User) -> None:
-    page = notion.create_page(parent=root_page, title='Test Local Remote Mention Block CMP')
+    # Test user mention
+    page = notion.create_page(parent=root_page, title='Test Local Remote User Mention Block CMP')
     user_ref = UserRef(id=person.id)
     mention_obj = objs.MentionUser.build_mention_from(user=user_ref, style=objs.Annotations())
+    mention_text = Text.wrap_obj_ref(obj_refs=[mention_obj])
+
+    block_child = uno.Paragraph(text=mention_text)
+    another_block_child = uno.Paragraph(text=mention_text)
+    assert block_child == another_block_child
+
+    page.append(blocks=[block_child])
+
+    page.reload()
+    assert len(page.children) == 1
+    assert page.children[0] == block_child
+    assert page.children[0] == another_block_child
+
+    # Test page mention
+    page = notion.create_page(parent=root_page, title='Test Local Remote Page Mention Block CMP')
+    page_obj_ref = ObjectRef(id=root_page.id)
+    mention_obj = objs.MentionPage.build_mention_from(page=page_obj_ref, style=objs.Annotations())
     mention_text = Text.wrap_obj_ref(obj_refs=[mention_obj])
 
     block_child = uno.Paragraph(text=mention_text)